Refactor : Changed files structure and added uuid package
