+++ src/Components/Cart/index.js
@@ -1,0 +1,1 @@
+export { Cart } from "./Cart";
