+++ src/Components/Products/ProductItem.jsx
@@ -1,5 +1,5 @@
-import { useData } from "../Context";
-import { Actions } from "../Context";
+// import axios from "axios";
+import { Actions,useData } from "../../Context";
 export const ProductItem = ({ product }) => {
   const { dispatch, setToast, setToastMessage, state } = useData();
   const inCartProducts = (id) => {
@@ -23,9 +23,11 @@ export const ProductItem = ({ product }) => {
 
         {inWishProducts(product.id) ? (
           <span
-            onClick={() => {
+            onClick={async() => {
               setToast("true");
               setToastMessage(`${product.brandName} removed from  wishlist`);
+              // const response = await axios.delete(`/api/wishes/${product.id}`);
+              // console.log({response})
               dispatch({
                 type: Actions.REMOVE_FROM_WISHLIST,
                 payload: product.id,
@@ -38,9 +40,13 @@ export const ProductItem = ({ product }) => {
           </span>
         ) : (
           <span
-            onClick={() => {
+            onClick={async () => {
               setToast("true");
               setToastMessage(`${product.brandName} is added to wishlist`);
+              // const {data:wish} = await axios.post(`/api/wishes`,
+              // {wish:{...product,count:1}}
+              // );
+              // console.log({wish});
               dispatch({ type: Actions.ADD_TO_WISHLIST, payload: product.id });
             }}
             className="wish-button"
@@ -52,9 +58,11 @@ export const ProductItem = ({ product }) => {
 
         {inCartProducts(product.id) ? (
           <button
-            onClick={() => {
+            onClick={async () => {
               setToast("true");
-              setToastMessage(`${product.brandName} is added to Cart`);
+              setToastMessage(`${product.brandName} is removed from Cart`);
+              // const response = await axios.delete(`/api/carts/${product.id}`);
+              // console.log({response})
               dispatch({ type: Actions.REMOVE_FROM_CART, payload: product.id });
             }}
             className="button button-primary"
@@ -63,9 +71,13 @@ export const ProductItem = ({ product }) => {
           </button>
         ) : (
           <button
-            onClick={() => {
+            onClick={async() => {
               setToast("true");
               setToastMessage(`${product.brandName} is added to Cart`);
+              // const {data:cart} = await axios.post(`/api/carts`,
+              // {cart:{...product,count:1}}
+              // );
+              // console.log(cart)
               dispatch({ type: Actions.ADD_TO_CART, payload: product.id });
             }}
             className="button button-primary"
