+++ src/Components/Products/index.js
@@ -1,0 +1,1 @@
+export { Products } from "./Products";
