+++ src/Components/WishList/WishListItem.jsx
@@ -1,4 +1,5 @@
-import { useData, Actions } from "../Context";
+import { useData, Actions } from "../../Context";
+// import axios from "axios";
 export const WishListItem = ({ product }) => {
     const { dispatch, setToast, setToastMessage } = useData();
     return (
@@ -10,9 +11,11 @@ export const WishListItem = ({ product }) => {
           <p className="card-desc">Count:{product.count}</p>
           <p className="card-desc">Rs {product.price}</p>
           <button
-            onClick={() => {
+            onClick={async() => {
               setToast("true");
               setToastMessage(`${product.brandName} removed from wish list`);
+              // const response = await axios.delete(`/api/wishes/${product.id}`);
+              // console.log({response})
               dispatch({
                 type: Actions.REMOVE_FROM_WISHLIST,
                 payload: product.id,
@@ -35,37 +38,38 @@ export const WishListItem = ({ product }) => {
           >
             Move To Cart
           </button>
-          <div
-            style={{
-              display: "flex",
-              justifyContent: "center",
-              alignItems: "center",
-            }}
-          >
-            <button
-              className="button button-border border-primary"
-              onClick={() => {
-                dispatch({
-                  type: Actions.INCREASE_ITEM_IN_WISHLIST,
-                  payload: product.id,
-                });
-              }}
-            >
-              +
-            </button>
-            <button
-              className="button button-border border-primary"
-              onClick={() => {
-                dispatch({
-                  type: Actions.DECREASE_ITEM_IN_WISHLIST,
-                  payload: product.id,
-                });
-              }}
-            >
-              -
-            </button>
-          </div>
         </div>
       </div>
     );
   };
+
+  // <div
+  //           style={{
+  //             display: "flex",
+  //             justifyContent: "center",
+  //             alignItems: "center",
+  //           }}
+  //         >
+  //           <button
+  //             className="button button-border border-primary"
+  //             onClick={() => {
+  //               dispatch({
+  //                 type: Actions.INCREASE_ITEM_IN_WISHLIST,
+  //                 payload: product.id,
+  //               });
+  //             }}
+  //           >
+  //             +
+  //           </button>
+  //           <button
+  //             className="button button-border border-primary"
+  //             onClick={() => {
+  //               dispatch({
+  //                 type: Actions.DECREASE_ITEM_IN_WISHLIST,
+  //                 payload: product.id,
+  //               });
+  //             }}
+  //           >
+  //             -
+  //           </button>
+  //         </div>
