+++ src/Components/WishList/index.js
@@ -1,0 +1,1 @@
+export { WishList } from "./WishList";
