+++ src/Context/Products.js
@@ -1,6 +1,7 @@
+import { v4 } from "uuid";
 export const products = [
   {
-    id: 0,
+    id: v4(),
     url:
       "https://assets.myntassets.com/h_1440,q_90,w_1080/v1/assets/images/4318138/2018/5/4/11525433792765-HERENOW-Men-Black-Printed-Round-Neck-T-shirt-2881525433792598-1.jpg",
     brandName: "Here and Now",
@@ -10,7 +11,7 @@ export const products = [
     inStock: false,
   },
   {
-    id: 1,
+    id: v4(),
     url:
       "https://assets.myntassets.com/h_1440,q_90,w_1080/v1/assets/images/1700944/2019/6/8/972c9498-3a37-4d5d-976c-4493b4d5c0021559989322793-HRX-by-Hrithik-Roshan-Men-Yellow-Printed-Round-Neck-T-Shirt--1.jpg",
     brandName: "HRX by Hrithik",
@@ -20,7 +21,7 @@ export const products = [
     inStock: true,
   },
   {
-    id: 2,
+    id: v4(),
     url:
       "https://assets.myntassets.com/h_1440,q_90,w_1080/v1/assets/images/2312468/2018/2/21/11519195992929-Roadster-Women-Maroon-Solid-Round-Neck-T-shirt-7951519195992737-1.jpg",
     brandName: "Roadster",
@@ -30,7 +31,7 @@ export const products = [
     inStock: true,
   },
   {
-    id: 3,
+    id: v4(),
     url:
       "https://assets.myntassets.com/h_1440,q_90,w_1080/v1/assets/images/11900232/2021/2/15/772d4a99-5ea0-40ac-bf21-4c4854708a491613387487527-ether-Men-Red--Navy-Blue-Pure-Cotton-Striped-Polo-Collar-T-s-1.jpg",
     brandName: "Ether",
@@ -40,7 +41,7 @@ export const products = [
     inStock: false,
   },
   {
-    id: 4,
+    id: v4(),
     url:
       "https://assets.myntassets.com/h_1440,q_90,w_1080/v1/assets/images/13420828/2021/2/18/61a79dc5-805d-4c2f-a5c1-996a78742f531613633242784-US-Polo-Assn-Men-Tshirts-7961613633241342-1.jpg",
     brandName: "US Polo Assn.",
@@ -50,7 +51,7 @@ export const products = [
     inStock: true,
   },
   {
-    id: 5,
+    id: v4(),
     url:
       "https://assets.myntassets.com/h_1440,q_90,w_1080/v1/assets/images/11067734/2019/12/10/2c445402-038d-4e7d-b07f-62284a09f4a01575972547961-DILLINGER-Men-Tshirts-4071575972546110-3.jpg",
     brandName: "Dillinger",
