+++ src/api/mockServer.js
@@ -1,8 +1,6 @@
 import { createServer, Model, RestSerializer } from "miragejs";
 import { products } from "../Context";
 export default function mockServer() {
-  const wishListProducts = [];
-  const cartProducts = [];
   createServer({
     serializers: {
       application: RestSerializer,
@@ -10,34 +8,27 @@ export default function mockServer() {
 
     models: {
       product: Model,
-      wishList: Model,
-      cartList: Model,
+      wish: Model,
+      cart: Model,
+    },
+    seeds(server) {
+      products.forEach((item) => {
+        server.create("product", item);
+      });
+      [].forEach((item) => {
+        server.create("wish", item);
+      });
+      [].forEach((item) => {
+        server.create("cart", item);
+      });
     },
 
     routes() {
       this.namespace = "api";
       this.timing = 1000;
       this.resource("products");
-      this.resource("wishList");
-      this.resource("cartList");
-      this.get("/products", () => {
-        return {
-          products,
-        };
-      });
-      this.get("/wishList", () => {
-        return {
-          wishListProducts,
-        };
-      });
-      this.post("/wishList", (request) => {
-        console.log(request);
-      });
-      this.get("/cartList", () => {
-        return {
-          cartProducts,
-        };
-      });
+      this.resource("wishes");
+      this.resource("carts");
     },
   });
 }
